BC-2331 - vue3 layouts (#2280)
* adds a default layout including logged in and logged out state
* moves auth to DefaultLayout component
* correct requireAuth to requireLogin
