+++ vue3_poc/src/App.vue
@@ -1,31 +1,16 @@
 <template>
-	<div class="mx-auto w-50">
-		<div class="text-right">
-			<v-btn v-if="isLoggedIn" @click="onLogout">Logout</v-btn>
-		</div>
+	<default-layout>
 		<router-view />
-	</div>
+	</default-layout>
 </template>
 
 <script>
-import { computed, defineComponent } from "vue";
-import { useAuthStore } from "./store/auth";
+import { defineComponent } from "vue";
+import DefaultLayout from "@/layouts/DefaultLayout.vue";
 
 export default defineComponent({
-	setup() {
-		const authStore = useAuthStore();
-
-		const onLogout = () => {
-			authStore.logout();
-			window.location.assign("/logout");
-		};
-
-		const isLoggedIn = computed(() => authStore.isLoggedIn);
-
-		return {
-			onLogout,
-			isLoggedIn,
-		};
+	components: {
+		DefaultLayout,
 	},
 });
 </script>
